requirements.txt
grpcio
scikit-learn
uvicorn[standard]
fastapi
protobuf2pydantic
click==7.1.1
requests